added sprite compatability. some other uhhh things components too

+++ Source/Engine/Core/MathUtils.h
@@ -3,9 +3,9 @@
 #include <utility>
 
 namespace kiko {
-	constexpr float Pi = 3.14159265359f;
-	constexpr float TwoPi = 6.28318530718;
-	constexpr float HalfPi = 1.57079632679;
+	constexpr float Pi = (float)3.14159265359f;
+	constexpr float TwoPi = (float)6.28318530718;
+	constexpr float HalfPi = (float)1.57079632679;
 
 	constexpr float RadiansToDegrees(float radians) { return radians * (180.0f / Pi); }
 	constexpr float DegreesToRadians(float degrees) { return degrees * (Pi / 180.0f); }
