+++ Source/Game/Game/SpaceGame.h
@@ -29,10 +29,10 @@ public:
 
 private:
 	eState m_state = eState::Title;
-	float m_spawnTimer = 0;
+	float m_spawnTimer = 0.0f;
 	float m_spawnTime = 3.0f;
 
-	int m_stateTimer = 0;
+	float m_stateTimer = 0.0f;
 
 	std::shared_ptr<kiko::Font> m_font;
 	std::unique_ptr<kiko::Text> m_scoreText;
